Solve problem 2.3 - test and fix console interface
